feat: Use long-click on list for accessibility swipe
Previously the longClick was used to launch a ContextMenu which provided
the option to remove that single item from its list. Other options in
this ContextMenu were duplicates found in the AppBar menu. Providing a
swipe-equivalent is more valuable than what the ContextMenu provided.
The swipe option could have been added to the ContextMenu instead, but
that would significantly slow down those using accessibility.